Add configuration variables and modules for Express server
Here you can see we setup several const values. This is the first part of
the server configuration file. In the first 3 lines, we grab the express
module. Following that we assign the express() function to the app const
to be able to use throughout our application.

BodyParser allows us to import a module that we use to do things like parse
JSON we get back from our requests.

The environment variable defaults to development if no environment variable
is found. This then is used to load the knex (database) configuration we need
from the knex database setup file. Finally, our database is loaded into the
application and this is the const we will later call our insert, update, destroy,
etc. functions on to interact with the database.