Add further configuration and root route to server file
We use the bodyParser module to automatically parse JSON into JS objects from
incoming responses from our requests. We set the port to 3000 if it is otherwise
not set in the environment variables. Finally, we listen continuously to check
for incoming connections to any of the routes listed above line 18.