Add migration for back end database
Here I've added the complete migration for our back end database. We have the
songs table with its attributes, the playlist table with its attributes, and
a joins table to combine the two tables when necessary.

Knex migrations require both the up (adding to database) and the down(removing
from database) in its migrations unlike the ones we used in Rails with AR.