Add POST function for /api/v1/songs
This closes issues #4 and #5. This function takes the POST request and parses
the JSON body into an object. We iterate over the object to make sure all fields
are there and correctly sent. We also check that the song rating field is
only a number between 1 and 100.

After checks have been completed, we insert the song into the database. We only
want to return the id, name, artist name, genre, and song rating from the
database, we do not need the timestamps. Then we respond with a content
created code and the newly created song object from our database.