Add post action for playlists
Simple function for POST action for creating a playlist. This just
specifies its name in the database so it can be further used to
add songs to separately. The only requirement is that the playlist_name
property is there and filled. If the requirement is met, we go ahead and
add the playlist to our database.