Add export for testing, test environment set
In our server.js file, I've added the export on line 67 to make sure we
can export the bulk of the app commands so we can run the testing suite
on them.

In the testing file, I have set the environment variable to test so
that it uses the test database instead of the development environment
database.