Refactor POST song to playlist
Refactored the POST song to playlist function by making the then chain
off of the first promise call, not nested down the line. This enabled
me to write an IF statement around the playlist_songs insert so that
it will only hit this new knex promise if songName and playlistName
are found in the database. All tests passing!